refactor: change folders names from plural to singular
