+++ src/main/java/ucsal/cauzy/domain/service/LogAcoesService.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ucsal.cauzy.domain.entity.LogAcoes;
 import ucsal.cauzy.domain.repository.LogAcoesRepository;
-import ucsal.cauzy.domain.utils.exceptions.ResourceNotFoundException;
+import ucsal.cauzy.domain.utils.exception.ResourceNotFoundException;
 import ucsal.cauzy.rest.dto.LogAcoesDTO;
 import ucsal.cauzy.rest.mapper.LogAcoesMapper;
