+++ src/main/java/ucsal/cauzy/domain/utils/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package ucsal.cauzy.domain.utils.exceptions;
+package ucsal.cauzy.domain.utils.exception;
 
 public class NotFoundException extends RuntimeException{
     public NotFoundException (String message){
